Removed texts files

+++ src/map/elemental.cpp
@@ -865,9 +865,9 @@ uint64 ElementalDatabase::parseBodyNode(const YAML::Node &node) {
 	}
 
 	if (this->nodeExists(node, "Defense")) {
-		uint16 def;
+		int32 def;
 
-		if (!this->asUInt16(node, "Defense", def))
+		if (!this->asInt32(node, "Defense", def))
 			return 0;
 
 		if (def < DEFTYPE_MIN || def > DEFTYPE_MAX) {
@@ -882,9 +882,9 @@ uint64 ElementalDatabase::parseBodyNode(const YAML::Node &node) {
 	}
 
 	if (this->nodeExists(node, "MagicDefense")) {
-		uint16 def;
+		int32 def;
 
-		if (!this->asUInt16(node, "MagicDefense", def))
+		if (!this->asInt32(node, "MagicDefense", def))
 			return 0;
 
 		if (def < DEFTYPE_MIN || def > DEFTYPE_MAX) {
